Resolved issue with negative progress percentage due to use of int for content length and happen to exceed max int value.

+++ src/main/java/tz/co/nezatech/dev/nezahttp/HttpPostProgressListener.java
@@ -1,6 +1,6 @@
 package tz.co.nezatech.dev.nezahttp;
 
 public interface HttpPostProgressListener {
-	public void progressChanged(int bytesAdded, int currentProgress, int totalSize);
-	public void postCompleted(int totalSize);
+	public void progressChanged(long bytesAdded, long currentProgress, long totalSize);
+	public void postCompleted(long totalSize);
 }
